Added tests and changed logout and password helper
