+++ pages/product.py
@@ -3,8 +3,7 @@ Product Page comes here.
 """
 import logging
 
-from selenium.webdriver.support import expected_conditions as EC
-from selenium.webdriver.support.wait import WebDriverWait
+import time
 
 from locators.product import ProductPageLocators
 from .base import BasePage
@@ -35,8 +34,6 @@ class ProductPage(BasePage):
         :return: Cart Page Object.
         """
         logging.info('clicking Cart Link in top Bar')
-        WebDriverWait(self.driver, 10).until(
-            EC.element_to_be_clickable(ProductPageLocators.GO_CART)
-        )
+        time.sleep(1)
         self.driver.find_element(*ProductPageLocators.GO_CART).click()
         return CartPage(self.driver)
